Fix a bunch of clippy lints.

+++ src/convert/mod.rs
@@ -253,8 +253,7 @@ pub(crate) fn generate_conversion_array(rf_inputs: &[Rfinput]) -> Vec<LegacyConv
     // Create an output vector so we can lookup where to get data from the legacy HDU, given a baseline/ant1/ant2
     let baseline_count = get_baseline_count(128);
 
-    let mut conversion_table: Vec<LegacyConversionBaseline> =
-        Vec::with_capacity(baseline_count as usize);
+    let mut conversion_table: Vec<LegacyConversionBaseline> = Vec::with_capacity(baseline_count);
 
     // Our row tile and column tile.  Now 2 pols each so only 128 in legacy obs
     for row_tile in 0..128 {
@@ -276,8 +275,8 @@ pub(crate) fn generate_conversion_array(rf_inputs: &[Rfinput]) -> Vec<LegacyConv
     }
 
     // Ensure we processed all baselines
-    assert_eq!(baseline, baseline_count as usize);
-    assert_eq!(conversion_table.len(), baseline_count as usize);
+    assert_eq!(baseline, baseline_count);
+    assert_eq!(conversion_table.len(), baseline_count);
 
     trace!("legacy_conversion_table: {:?}", conversion_table);
